Added Bash files for compiling and running.
For UNIX systems: From the src directory and in terminal: chmod 777
musicftp-[server/client]
Windows users use the equivalent of chmod. 

You will only have to do that once. After that, it will compile and
start the client or server.For me, I can click and it and run it. It
builds all files then runs the program.